refactor(RadioButtonGroup): more opinionated radio button group abstraction
